socialapi: move tests to their own files
