+++ go/src/socialapi/workers/algoliaconnector/algoliaconnector/algoliaconnector_test.go
@@ -19,182 +19,6 @@ import (
 
 const TestTimeout = 6 * time.Minute
 
-func TestTopicSaved(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	Convey("given some fake topic channel", t, func() {
-		mockTopic := models.NewChannel()
-		mockTopic.TypeConstant = models.Channel_TYPE_TOPIC
-		Convey("it should save the document to algolia", func() {
-			err := handler.TopicSaved(mockTopic)
-			So(err, ShouldBeNil)
-		})
-	})
-	Convey("given some fake non-topic channel", t, func() {
-		mockTopic := models.NewChannel()
-		mockTopic.TypeConstant = models.Channel_TYPE_PRIVATE_MESSAGE
-		Convey("it should save the document to algolia", func() {
-			err := handler.TopicSaved(mockTopic)
-			So(err, ShouldBeNil)
-		})
-	})
-}
-
-func TestTopicUpdated(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	rand.Seed(time.Now().UnixNano())
-
-	appConfig := config.MustRead(runner.Conf.Path)
-	modelhelper.Initialize(appConfig.Mongo)
-	defer modelhelper.Close()
-
-	Convey("given some fake topic channel", t, func() {
-		mockTopic := models.NewChannel()
-		mockTopic.Id = rand.Int63()
-		mockTopic.TypeConstant = models.Channel_TYPE_TOPIC
-		Convey("it should save the document to algolia", func() {
-			err := handler.TopicSaved(mockTopic)
-			So(err, ShouldBeNil)
-			err = makeSureTopic(handler, mockTopic.Id, func(record map[string]interface{}, err error) bool {
-				if err != nil {
-					return false
-				}
-
-				return true
-			})
-
-			So(err, ShouldBeNil)
-
-			Convey("given some existing topic channel", func() {
-				mockTopic.TypeConstant = models.Channel_TYPE_LINKED_TOPIC
-				Convey("it should be able to remove it", func() {
-					err := handler.TopicUpdated(mockTopic)
-					So(err, ShouldBeNil)
-
-					err = makeSureTopic(handler, mockTopic.Id, func(record map[string]interface{}, err error) bool {
-						if IsAlgoliaError(err, ErrAlgoliaObjectIdNotFoundMsg) {
-							return true
-						}
-
-						return false
-					})
-
-					So(err, ShouldBeNil)
-
-					Convey("removing a deleted channel should return success", func() {
-						err := handler.TopicUpdated(mockTopic)
-						So(err, ShouldBeNil)
-					})
-				})
-			})
-			Convey("removing a non-existing channel should return success", func() {
-				mockTopic.Id++
-				err := handler.TopicUpdated(mockTopic)
-				So(err, ShouldBeNil)
-			})
-		})
-	})
-}
-
-func TestAccountSaved(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	Convey("given some fake account", t, func() {
-		mockAccount := &models.Account{
-			OldId:   bson.NewObjectId().Hex(),
-			Id:      100000000,
-			Nick:    "fake-nickname",
-			IsTroll: false,
-		}
-		Convey("it should save the document to algolia", func() {
-			err := handler.AccountSaved(mockAccount)
-			So(err, ShouldBeNil)
-		})
-	})
-}
-
-func TestMessageListSaved(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	Convey("messages can be saved", t, func() {
-		mockMessage, _ := createAndSaveMessage()
-		mockListing := getListings(mockMessage)[0]
-
-		So(handler.MessageListSaved(&mockListing), ShouldBeNil)
-		So(doBasicTestForMessage(handler, mockListing.MessageId), ShouldBeNil)
-	})
-
-	Convey("messages can be cross-indexed", t, func() {
-		mockMessage, owner := createAndSaveMessage()
-
-		// init channel
-		cm, err := createChannel(owner.Id)
-		So(err, ShouldBeNil)
-		So(cm, ShouldNotBeNil)
-		// init channel message list
-		cml := createChannelMessageList(cm.Id, mockMessage.Id)
-		So(cml, ShouldNotBeNil)
-
-		listings := getListings(mockMessage)
-		So(len(listings), ShouldEqual, 2)
-
-		So(handler.MessageListSaved(&listings[0]), ShouldBeNil)
-		So(doBasicTestForMessage(handler, listings[0].MessageId), ShouldBeNil)
-
-		So(handler.MessageListSaved(&listings[1]), ShouldBeNil)
-		err = makeSureMessage(handler, listings[1].MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			if len((record["_tags"]).([]interface{})) != 2 {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-	})
-}
-
-func doBasicTestForMessage(handler *Controller, id int64) error {
-	return makeSureMessage(handler, id, func(record map[string]interface{}, err error) bool {
-		if err != nil {
-			return false
-		}
-
-		if record == nil {
-			return false
-		}
-
-		return true
-	})
-}
-
-// makeSureMessage checks if the given id's get request returns the desired err, it
-// will re-try every 100ms until deadline of 15 seconds reached. Algolia doesnt
-// index the records right away, so try to go to a desired state
-func makeSureMessage(handler *Controller, id int64, f func(map[string]interface{}, error) bool) error {
-	deadLine := time.After(TestTimeout)
-	tick := time.Tick(time.Millisecond * 100)
-	for {
-		select {
-		case <-tick:
-			record, err := handler.get(IndexMessages, strconv.FormatInt(id, 10))
-			if f(record, err) {
-				return nil
-			}
-		case <-deadLine:
-			return errDeadline
-		}
-	}
-}
-
 // makeSureSynonyms checks if the given index's synonyms request returns the
 // desired err, it will re-try every 100ms until deadline of 15 seconds reached.
 // Algolia doesnt index the records right away, so try to go to a desired state
@@ -216,149 +40,6 @@ func makeSureSynonyms(handler *Controller, indexName string, f func([][]string, 
 			return errDeadline
 		}
 	}
-}
-
-// makeSureTopic checks if the given id's get request returns the desired err,
-// it will re-try every 100ms until deadline of 2 minutes. Algolia doesnt index
-// the records right away, so try to go to a desired state
-func makeSureTopic(handler *Controller, id int64, f func(map[string]interface{}, error) bool) error {
-	deadLine := time.After(TestTimeout)
-	tick := time.Tick(time.Millisecond * 100)
-	for {
-		select {
-		case <-tick:
-			record, err := handler.get(IndexTopics, strconv.FormatInt(id, 10))
-			if f(record, err) {
-				return nil
-			}
-		case <-deadLine:
-			return errDeadline
-		}
-	}
-}
-
-func TestMessageListDeleted(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	Convey("messages can be deleted", t, func() {
-		mockMessage, _ := createAndSaveMessage()
-		mockListing := getListings(mockMessage)[0]
-
-		So(handler.MessageListSaved(&mockListing), ShouldBeNil)
-		So(doBasicTestForMessage(handler, mockListing.MessageId), ShouldBeNil)
-
-		So(handler.MessageListDeleted(&mockListing), ShouldBeNil)
-		err := makeSureMessage(handler, mockListing.MessageId, func(record map[string]interface{}, err error) bool {
-			if err == nil {
-				return false
-			}
-
-			if record != nil {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-	})
-
-	Convey("cross-indexed messages will not be deleted", t, func() {
-		mockMessage, owner := createAndSaveMessage()
-
-		// init channel
-		cm, err := createChannel(owner.Id)
-		So(err, ShouldBeNil)
-		So(cm, ShouldNotBeNil)
-		// init channel message list
-		cml := createChannelMessageList(cm.Id, mockMessage.Id)
-		So(cml, ShouldNotBeNil)
-
-		listings := getListings(mockMessage)
-		So(len(listings), ShouldEqual, 2)
-
-		So(handler.MessageListSaved(&listings[0]), ShouldBeNil)
-		err = makeSureMessage(handler, listings[0].MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			if len((record["_tags"]).([]interface{})) != 1 {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-
-		So(handler.MessageListSaved(&listings[1]), ShouldBeNil)
-
-		err = makeSureMessage(handler, listings[1].MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			if len((record["_tags"]).([]interface{})) != 2 {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-
-		So(handler.MessageListDeleted(&listings[1]), ShouldBeNil)
-
-		err = makeSureMessage(handler, listings[1].MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			if len((record["_tags"]).([]interface{})) != 1 {
-				return false
-			}
-
-			return true
-		})
-
-		So(err, ShouldBeNil)
-	})
-}
-
-func TestMessageUpdated(t *testing.T) {
-	runner, handler := getTestHandler()
-	defer runner.Close()
-
-	Convey("messages can be updated", t, func() {
-		mockMessage, _ := createAndSaveMessage()
-		mockListing := getListings(mockMessage)[0]
-
-		So(handler.MessageListSaved(&mockListing), ShouldBeNil)
-		err := makeSureMessage(handler, mockListing.MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-
-		mockMessage.Body = "updated body"
-
-		So(mockMessage.Update(), ShouldBeNil)
-		So(handler.MessageUpdated(mockMessage), ShouldBeNil)
-		err = makeSureMessage(handler, mockListing.MessageId, func(record map[string]interface{}, err error) bool {
-			if err != nil {
-				return false
-			}
-
-			if record["body"].(string) != "updated body" {
-				return false
-			}
-
-			return true
-		})
-		So(err, ShouldBeNil)
-	})
 }
 
 func TestIndexSettings(t *testing.T) {
@@ -395,7 +76,7 @@ func TestIndexSettings(t *testing.T) {
 
 				So(cl.Create(), ShouldBeNil)
 
-				So(handler.CreateSynonym(cl), ShouldBeNil)
+				So(handler.ChannelLinkCreated(cl), ShouldBeNil)
 
 				Convey("synonyms should be in settings", func() {
 					err = makeSureSynonyms(handler, IndexMessages, func(synonyms [][]string, err error) bool {
@@ -437,7 +118,7 @@ func TestIndexSettings(t *testing.T) {
 						}
 
 						So(cl.Create(), ShouldBeNil)
-						So(handler.CreateSynonym(cl), ShouldBeNil)
+						So(handler.ChannelLinkCreated(cl), ShouldBeNil)
 
 						Convey("synonyms should be in settings", func() {
 							err = makeSureSynonyms(handler, IndexMessages, func(synonyms [][]string, err error) bool {
@@ -489,7 +170,7 @@ func TestIndexSettings(t *testing.T) {
 
 							So(cl.Create(), ShouldBeNil)
 
-							So(handler.CreateSynonym(cl), ShouldBeNil)
+							So(handler.ChannelLinkCreated(cl), ShouldBeNil)
 
 							Convey("all synonyms should be in settings", func() {
 								err = makeSureSynonyms(handler, IndexMessages, func(synonyms [][]string, err error) bool {
